src/evaluate.py
import torch
from .preprocess import get_data_loaders
from .model import get_model
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score

def evaluate_model():
    device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
    model = get_model()
    model.load_state_dict(torch.load('models/best_model.pth', map_location=device))
    model.to(device)
    model.eval()

    _, _, test_loader = get_data_loaders()

    all_preds = []
    all_labels = []
    with torch.no_grad():
        for inputs, labels in test_loader:
            inputs, labels = inputs.to(device), labels.to(device)
            outputs = model(inputs)
            _, predicted = torch.max(outputs, 1)
            all_preds.extend(predicted.cpu().numpy())
            all_labels.extend(labels.cpu().numpy())

    acc = accuracy_score(all_labels, all_preds)
    prec = precision_score(all_labels, all_preds)
    rec = recall_score(all_labels, all_preds)
    f1 = f1_score(all_labels, all_preds)

    print(f'Accuracy: {acc:.4f}')
    print(f'Precision: {prec:.4f}')
    print(f'Recall: {rec:.4f}')
    print(f'F1 Score: {f1:.4f}')

if __name__ == '__main__':
    evaluate_model()