requirements.txt
torch==2.11.0
torchvision==0.26.0
streamlit
Pillow
requests
scikit-learn
matplotlib